update runner to config project vitest
